Change background colors for cells at game end
Change background colors for gameboard cells when the game ends such that:
A. If either player has won, the cells in the winning line have a background color matching that of the winning player, with `35%` opacity.
B. If neither player has won, the cells in the winning line have the background color used for a draw, with `20%` opacity.